ref(http-client): proper class files and minor fixes
